Bug#45288: pb2 returns a lot of compilation warnings

+++ sql/rpl_slave.cc
@@ -2446,11 +2446,11 @@ static void sql_delay_event(Log_event *ev, THD *thd, Relay_log_info *rli)
                         "ev->when= %lu "
                         "rli->mi->clock_diff_with_master= %lu "
                         "now= %ld "
-                        "sql_delay_end= %lu "
+                        "sql_delay_end= %ld "
                         "nap_time= %ld",
                         sql_delay, (long)ev->when,
                         rli->mi->clock_diff_with_master,
-                        (long)now, sql_delay_end, (long)nap_time));
+                        (long)now, (long)sql_delay_end, (long)nap_time));
 
     if (sql_delay_end > now)
     {
